refactor: migrate messenger to ts
