+++ src/types/react-transition-group.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-transition-group';
